A bit of a rearrangement.
