+++ tests/interactive/test_network.py
@@ -1,5 +1,3 @@
-import asyncio
-import os
 import socket
 from collections.abc import Callable, Awaitable
 
@@ -12,7 +10,6 @@ import network
 from configuration import NODE_NAME, NODE_ROLE
 from interactive import configuration
 from network import NetworkController, HEADER_NAME, HEADER_ROLE
-from polyfills import cpu
 from runner import Runner
 
 
@@ -295,276 +292,6 @@ def mock_send_message(path: str, host: str = configuration.NODE_COORDINATOR,
                       data=None, json=None):
     pass
     # TODO: This needs to return a valid response object.
-
-
-class TestHttpRoutes:
-    """
-    NOTE: Even though these tests all specify (or at least try to) the correct route
-          in the TestRequest(), those routes are not typically needed for the test methods
-          as they expect to have been routed to correctly anyway.
-    """
-
-    @pytest.fixture(autouse=True)
-    def send_message_patched(self, monkeypatch):
-        monkeypatch.setattr(network, 'send_message', mock_send_message)
-
-    def test_index_returns_index_html_with_get(self) -> None:
-        """
-        Validates that the index.html file is returned with no additional headers.
-        """
-        validate_methods({GET}, "/", network.index)
-
-        path = os.getcwd()
-        try:
-            dir_path = os.path.dirname(os.path.realpath(__file__))
-            os.chdir(os.path.join(dir_path, "../.."))
-            request = TestRequest(GET, "/")
-            response = network.index(request)
-            assert response._filename == "index.html"
-            assert response._root_path == 'interactive/html'
-            assert response._status == OK_200
-            assert len(response._headers) == 0
-        finally:
-            os.chdir(path)
-
-    def test_cpu_information(self) -> None:
-        """
-        Validates that the cpu information is returned with no additional headers.
-        """
-        validate_methods({GET}, "/cpu-information", network.cpu_information)
-
-        request = TestRequest("GET", "/cpu-information")
-        response = network.cpu_information(request)
-        assert response._data == cpu.info()
-        assert response._status == OK_200
-        assert len(response._headers) == 0
-
-    def test_inspect(self) -> None:
-        """
-        Validates that the inspect web page is returned with no additional headers.
-        """
-        validate_methods({GET}, "/inspect", network.inspect)
-
-        request = TestRequest(GET, "/inspect")
-        response = network.inspect(request)
-        assert response._body == "TODO inspect"
-        assert response._status == OK_200
-        assert len(response._headers) == 0
-
-    def test_register(self) -> None:
-        validate_methods({GET, POST, PUT}, "/register", network.register)
-        # TODO: Write the rest of the tests
-
-    def test_unregister(self) -> None:
-        validate_methods({GET, POST, PUT}, "/unregister", network.unregister)
-        # TODO: Write the rest of the tests
-
-    def test_heartbeat(self) -> None:
-        validate_methods({GET, POST, PUT}, "/heartbeat", network.heartbeat)
-        # TODO: Write the rest of the tests
-
-    def test_restart(self, monkeypatch) -> None:
-        """
-        Validates that restart returns with no additional headers. The restart
-        won't actually restart a Desktop PC as the polyfill is a noop; though
-        we monkeypatch the restart() function anyway.
-        """
-        validate_methods({GET}, "/restart", network.restart)
-
-        restart_called_count = 0
-
-        def test_restart_fn():
-            nonlocal restart_called_count
-            restart_called_count += 1
-
-        monkeypatch.setattr(network, 'cpu_restart', test_restart_fn)
-
-        # Create an event loop which is needed for the async restart request.
-        request = TestRequest(GET, "/restart")
-
-        async def __execute():
-            # Before calling the request, there should only be a single
-            # async task which is this one.
-            assert len(asyncio.all_tasks()) == 1
-            response = network.restart(request)
-            assert response._body == network.YES
-            assert response._status == OK_200
-
-            # There should now be 2 tasks, this one and the restart async task.
-            assert len(asyncio.all_tasks()) == 2
-
-            # Await for long enough for the restart to get executed.
-            while len(asyncio.all_tasks()) > 1:
-                await asyncio.sleep(0.1)
-
-        asyncio.run(__execute())
-        assert restart_called_count == 1
-
-    def test_alive(self) -> None:
-        """
-        Validates that alive returns with no additional headers.
-        """
-        validate_methods({GET}, "/alive", network.alive)
-
-        request = TestRequest(GET, "/alive")
-        response = network.alive(request)
-        assert response._body == network.YES
-        assert response._status == OK_200
-
-    def test_name(self) -> None:
-        """
-        Validates that name returns with no additional headers.
-        """
-        validate_methods({GET}, "/name", network.name)
-
-        request = TestRequest(GET, "/name")
-        response = network.name(request)
-        assert response._body == configuration.NODE_NAME
-        assert response._status == OK_200
-
-    def test_role(self) -> None:
-        """
-        Validates that role returns with no additional headers.
-        """
-        validate_methods({GET}, "/role", network.role)
-
-        request = TestRequest(GET, "/role")
-        response = network.role(request)
-        assert response._body == configuration.NODE_ROLE
-        assert response._status == OK_200
-
-    def test_details(self) -> None:
-        """
-        Validates that the details information is returned with no additional headers.
-        """
-        validate_methods({GET}, "/details", network.details)
-
-        request = TestRequest(GET, "/details")
-        response = network.details(request)
-        assert response._data == configuration.details()
-        assert response._status == OK_200
-        assert len(response._headers) == 0
-
-    def test_led_blink(self, monkeypatch) -> None:
-        """
-        Validates the led_blink route calls the appropriate receive function.
-        """
-        validate_methods({GET}, "/led/blink", network.led_blink)
-
-        message_called_count = 0
-
-        def test_message_fn(r: Request) -> str:
-            nonlocal message_called_count
-            message_called_count += 1
-            return 'LED has blinked'
-
-        monkeypatch.setattr(network, 'receive_blink_message', test_message_fn)
-
-        request = TestRequest(GET, "/led/blink")
-        response = network.led_blink(request)
-        assert response._body == 'LED has blinked'
-        assert response._status == OK_200
-        assert len(response._headers) == 0
-
-        assert message_called_count == 1
-
-    def test_led_state(self, monkeypatch) -> None:
-        """
-        Validates the led_state() returns the desired state..
-        """
-        validate_methods({GET}, "/led/<state>", network.led_state, "ON")
-
-        message_called_count = 0
-
-        def test_message_fn(request: Request, state: str) -> str:
-            nonlocal message_called_count
-            message_called_count += 1
-            return f'LED is {state}'
-
-        monkeypatch.setattr(network, 'receive_led_message', test_message_fn)
-
-        request = TestRequest(GET, "/led/<state>")
-        response = network.led_state(request, "ON")
-        assert response._body == 'LED is ON'
-        assert response._status == OK_200
-        assert len(response._headers) == 0
-
-        assert message_called_count == 1
-
-        response = network.led_state(request, "OFF")
-        assert response._body == 'LED is OFF'
-        assert response._status == OK_200
-        assert len(response._headers) == 0
-
-        assert message_called_count == 2
-
-    def test_lookup_all(self) -> None:
-        """
-        Validates that lookup_all() returns NOT_IMPLEMENTED_501 with no additional headers.
-        """
-        validate_methods({GET}, "/lookup/all", network.lookup_all)
-
-        request = TestRequest(GET, "/lookup/all")
-        response = network.lookup_all(request)
-        assert response._body == network.NO
-        assert response._status == NOT_IMPLEMENTED_501
-
-    def test_lookup_name(self) -> None:
-        """
-        Validates that lookup_name() returns NOT_IMPLEMENTED_501 with no additional headers.
-        """
-        validate_methods({GET}, "/lookup/name/<name>", network.lookup_name, "NAME")
-
-        request = TestRequest(GET, "/lookup/name/<name>")
-        response = network.lookup_name(request, "")
-        assert response._body == network.NO
-        assert response._status == NOT_IMPLEMENTED_501
-
-    def test_lookup_role(self) -> None:
-        """
-        Validates that lookup_role() returns NOT_IMPLEMENTED_501 with no additional headers.
-        """
-        validate_methods({GET}, "/lookup/role/<role>", network.lookup_role, "ROLE")
-
-        request = TestRequest(GET, "/lookup/role/<role>")
-        response = network.lookup_role(request, "")
-        assert response._body == network.NO
-        assert response._status == NOT_IMPLEMENTED_501
-
-
-class TestMessages:
-
-    @pytest.fixture(autouse=True)
-    def send_message_patched(self, monkeypatch):
-        monkeypatch.setattr(network, 'send_message', mock_send_message)
-
-    def test_send_message(self) -> None:
-        network.send_message("hi")
-        assert False
-
-    def test_send_register_message(self) -> None:
-        assert False
-
-    def test_receive_register_message(self) -> None:
-        assert False
-
-    def test_send_unregister_message(self) -> None:
-        assert False
-
-    def test_receive_unregister_message(self) -> None:
-        assert False
-
-    def test_send_heartbeat_message(self) -> None:
-        assert False
-
-    def test_receive_heartbeat_message(self) -> None:
-        assert False
-
-    def test_receive_blink_message(self) -> None:
-        assert False
-
-    def test_receive_led_message(self) -> None:
-        assert False
 
 # print(request)
 # print(f"METHOD ... : '{request.method}'")
